chore(pages): move login and register pages to (auth) sub dir
