+++ app/(auth)/login/page.ts
@@ -1,1 +1,1 @@
-export { LoginPage as default } from '@/pages/login';
+export { LoginPage as default } from '@/pages/(auth)/login';
