+++ src/pages/(auth)/register/index.ts
