+++ src/pages/(auth)/register/ui/register.module.scss
