+++ src/pages/(auth)/register/ui/register.tsx
